requirements.txt
confluent_kafka
avro
fastavro
requests
faker
aws-msk-iam-sasl-signer-python